agent_runner/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
redis[hiredis]==5.2.0
anthropic==0.40.0
pydantic==2.10.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
slowapi==0.1.9
httpx==0.28.0
python-dotenv==1.0.1
scrapling[fetchers]==0.2.9
